상태 수정
