+++ Assets/#1 Scripts/#0 State/PlayerOwnedStates.cs
@@ -2,22 +2,7 @@ using UnityEngine;
 
 namespace PlayerOwnedStates
 {
-    public class IsGround  : State<Player>
-    {
-        public override void Enter(Player entity)
-        {
-            
-        }
-        public override void Execute(Player entity)
-        {
-            
-        }
-        public override void Exit(Player entity)
-        {
-            
-        }
-    }
-    public class CanJump : State<Player>
+    public class IsMove : State<Player>
     {
         public override void Enter(Player entity)
         {
@@ -62,7 +47,7 @@ namespace PlayerOwnedStates
             
         }
     }
-    public class IsWall : State<Player>
+    public class CanDefence : State<Player>
     {
         public override void Enter(Player entity)
         {
@@ -77,7 +62,7 @@ namespace PlayerOwnedStates
             
         }
     }
-    public class IsSlope : State<Player>
+    public class IsDefencing : State<Player>
     {
         public override void Enter(Player entity)
         {
@@ -92,22 +77,7 @@ namespace PlayerOwnedStates
 
         }
     }
-    public class IsMove : State<Player>
-    {
-        public override void Enter(Player entity)
-        {
-            
-        }
-        public override void Execute(Player entity)
-        {
-            
-        }
-        public override void Exit(Player entity)
-        {
-            
-        }
-    }
-    public class IsFallAttacking : State<Player>
+    public class IsAttacking : State<Player>
     {
         public override void Enter(Player entity)
         {
