--- Assets/#1 Scripts/#1 Entity/Player/move.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: afc2babdae0bc6843b2de77edcb876dd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
