+++ Assets/#2 Prefabs/Player/player 1.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 52a11fbe8cd3b4becbe4007715b64acc
+guid: 7411f4a381fbfe0439086193039257bc
 PrefabImporter:
   externalObjects: {}
   userData:
